Bugfix: wyl files with multiple identical inputs
In the execution of wyl files, if you put two lines to execute with the
same sequence of characters, even if they should have evaluated to
different results, then the expected value for each would have been the
expected value for the last instance of that character sequence. This is
now fixed.